Initial LIS3MDL Driver Implementation
Description:
This commit introduces an initial implementation of the LIS3MDL driver for CubeIDE. It includes basic read and write functionality for sensor data and configuration registers. The code has been tested on the Nucleo-L412KB development board, and it is functional in our testing environment.

Known Issues:
- There is an issue with the conversion from raw sensor data to meaningful data. The current conversion logic may result in inaccurate or incorrect sensor readings. Further investigation and adjustment of the data conversion code are needed to address this issue.

Note: While the driver is currently tailored for the Nucleo-L412KB board, it can be made compatible with other boards by including the appropriate header file in the `.h` file.